Language.Haskell.GHC.Alter.Unexported.ByteCodeGen

Description

ByteCodeGen: Generate bytecode from Core

Documentation

byteCodeGen :: HscEnv -> Module -> CoreProgram -> [TyCon] -> Maybe ModBreaks -> IO CompiledByteCode

allocateTopStrings :: HscEnv -> [(Id, ByteString)] -> IO [(Var, RemotePtr ())]

coreExprToBCOs :: HscEnv -> Module -> CoreExpr -> IO UnlinkedBCO

simpleFreeVars :: CoreExpr -> AnnExpr Id DVarSet

type BCInstrList = OrdList BCInstr

newtype ByteOff

Constructors

ByteOff Int

newtype WordOff

Constructors

WordOff Int

wordsToBytes :: DynFlags -> WordOff -> ByteOff

bytesToWords :: DynFlags -> ByteOff -> WordOff

wordSize :: DynFlags -> ByteOff

type Sequel = ByteOff

type StackDepth = ByteOff

type BCEnv = Map Id StackDepth

Maps Ids to their stack depth. This allows us to avoid having to mess with it after each push/pop.

mkProtoBCO :: DynFlags -> name -> BCInstrList -> Either [AnnAlt Id DVarSet] (AnnExpr Id DVarSet) -> Int -> Word16 -> [StgWord] -> Bool -> [FFIInfo] -> ProtoBCO name

argBits :: DynFlags -> [ArgRep] -> [Bool]

schemeTopBind :: (Id, AnnExpr Id DVarSet) -> BcM (ProtoBCO Name)

schemeR :: [Id] -> (Id, AnnExpr Id DVarSet) -> BcM (ProtoBCO Name)

collect :: AnnExpr Id DVarSet -> ([Var], AnnExpr' Id DVarSet)

schemeR_wrk :: [Id] -> Id -> AnnExpr Id DVarSet -> ([Var], AnnExpr' Var DVarSet) -> BcM (ProtoBCO Name)

schemeER_wrk :: StackDepth -> BCEnv -> AnnExpr' Id DVarSet -> BcM BCInstrList

getVarOffSets :: DynFlags -> StackDepth -> BCEnv -> [Id] -> [(Id, Word16)]

truncIntegral16 :: Integral a => a -> Word16

trunc16W :: WordOff -> Word16

fvsToEnv :: BCEnv -> DVarSet -> [Id]

returnUnboxedAtom :: StackDepth -> Sequel -> BCEnv -> AnnExpr' Id DVarSet -> ArgRep -> BcM BCInstrList

schemeE :: StackDepth -> Sequel -> BCEnv -> AnnExpr' Id DVarSet -> BcM BCInstrList

schemeT :: StackDepth -> Sequel -> BCEnv -> AnnExpr' Id DVarSet -> BcM BCInstrList

mkConAppCode :: StackDepth -> Sequel -> BCEnv -> DataCon -> [AnnExpr' Id DVarSet] -> BcM BCInstrList

unboxedTupleReturn :: StackDepth -> Sequel -> BCEnv -> AnnExpr' Id DVarSet -> BcM BCInstrList

doTailCall :: StackDepth -> Sequel -> BCEnv -> Id -> [AnnExpr' Id DVarSet] -> BcM BCInstrList

findPushSeq :: [ArgRep] -> (BCInstr, Int, [ArgRep])

doCase :: StackDepth -> Sequel -> BCEnv -> AnnExpr Id DVarSet -> Id -> [AnnAlt Id DVarSet] -> Maybe Id -> BcM BCInstrList

generateCCall :: StackDepth -> Sequel -> BCEnv -> CCallSpec -> Id -> [AnnExpr' Id DVarSet] -> BcM BCInstrList

primRepToFFIType :: DynFlags -> PrimRep -> FFIType

mkDummyLiteral :: PrimRep -> Literal

maybe_getCCallReturnRep :: Type -> Maybe PrimRep

maybe_is_tagToEnum_call :: AnnExpr' Id DVarSet -> Maybe (AnnExpr' Id DVarSet, [Name])

implement_tagToId :: StackDepth -> Sequel -> BCEnv -> AnnExpr' Id DVarSet -> [Name] -> BcM BCInstrList

pushAtom :: StackDepth -> BCEnv -> AnnExpr' Id DVarSet -> BcM (BCInstrList, ByteOff)

mkMultiBranch :: Maybe Int -> [(Discr, BCInstrList)] -> BcM BCInstrList

data Discr

Constructors

DiscrI Int

DiscrW Word

DiscrF Float

DiscrD Double

DiscrP Word16

NoDiscr

lookupBCEnv_maybe :: Id -> BCEnv -> Maybe ByteOff

idSizeW :: DynFlags -> Id -> WordOff

idSizeB :: DynFlags -> Id -> ByteOff

bcIdArgRep :: Id -> ArgRep

bcIdPrimRep :: Id -> PrimRep

isFollowableArg :: ArgRep -> Bool

isVoidArg :: ArgRep -> Bool

multiValException :: a

isSupportedCConv :: CCallSpec -> Bool

Indicate if the calling convention is supported

unsupportedCConvException :: a

mkSlideB :: DynFlags -> ByteOff -> ByteOff -> OrdList BCInstr

mkSlideW :: Word16 -> WordOff -> OrdList BCInstr

splitApp :: AnnExpr' Var ann -> (AnnExpr' Var ann, [AnnExpr' Var ann])

bcView :: AnnExpr' Var ann -> Maybe (AnnExpr' Var ann)

isVAtom :: AnnExpr' Var ann -> Bool

atomPrimRep :: AnnExpr' Id ann -> PrimRep

atomRep :: AnnExpr' Id ann -> ArgRep

mkStackOffsets :: ByteOff -> [ByteOff] -> [ByteOff]

Let szsw be the sizes in bytes of some items pushed onto the stack, which has initial depth original_depth. Return the values which the stack environment should map these items to.

typeArgRep :: Type -> ArgRep

data BcM_State

Constructors

BcM_State

Fields

bcm_hsc_env :: HscEnv

uniqSupply :: UniqSupply

thisModule :: Module

nextlabel :: Word16

ffis :: [FFIInfo]

modBreaks :: Maybe ModBreaks

breakInfo :: IntMap CgBreakInfo

topStrings :: IdEnv (RemotePtr ())

newtype BcM r

Constructors

BcM (BcM_State -> IO (BcM_State, r))

ioToBc :: IO a -> BcM a

runBc :: HscEnv -> UniqSupply -> Module -> Maybe ModBreaks -> IdEnv (RemotePtr ()) -> BcM r -> IO (BcM_State, r)

thenBc :: BcM a -> (a -> BcM b) -> BcM b

thenBc_ :: BcM a -> BcM b -> BcM b

returnBc :: a -> BcM a

getHscEnv :: BcM HscEnv

emitBc :: ([FFIInfo] -> ProtoBCO Name) -> BcM (ProtoBCO Name)

recordFFIBc :: RemotePtr C_ffi_cif -> BcM ()

getLabelBc :: BcM Word16

getLabelsBc :: Word16 -> BcM [Word16]

getCCArray :: BcM (Array BreakIndex (RemotePtr CostCentre))

newBreakInfo :: BreakIndex -> CgBreakInfo -> BcM ()

newUnique :: BcM Unique

getCurrentModule :: BcM Module

getTopStrings :: BcM (IdEnv (RemotePtr ()))

newId :: Type -> BcM Id

tickFS :: FastString